requirements.txt
python-telegram-bot==13.1

requests==2.24.0

beautifulsoup4==4.9.1

python-dotenv>=0.10
